% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{chromosomeSlidingWindow}
\alias{chromosomeSlidingWindow}
\title{Run sliding window over chromosome for transposon readouts}
\usage{
chromosomeSlidingWindow(
  data,
  chromosome,
  windowStart,
  windowEnd,
  windowSize = 10000,
  stepSize = 1000,
  mode,
  numerator = NULL,
  denominator = NULL
)
}
\arguments{
\item{data}{Insertion site data (as formatted by TagMap pipeline).}

\item{chromosome}{Chromosome, formatted as "chr(number)".}

\item{windowStart}{Chromosome position (integer) to calculate window
coordinates from.}

\item{windowEnd}{Chromosome position (integer) to calculate window
coordinates to.}

\item{windowSize}{Size of each window, e.g. 1e3 or 1e4 bp long.}

\item{stepSize}{Size that each step increases, from minimal steps of 1 bp to
any other custom value.}

\item{mode}{Readout mode, eiter 'counts' for number of insertions per sliding
window or 'fraction' for number of insertions per sliding across experimental
conditions.}

\item{numerator}{Experimental condition(s) for numerator of fraction of
insertions (only required for fraction readout).}

\item{denominator}{Experimental condition(s) for denominator of fraction of
insertions (only required for fraction readout).}
}
\value{
Returns data table for putative insertion sites outputted by the
TagMap pipeline.
\describe{
\item{windowStart, windowEnd}{Chromosome start/end positions of a specific
window}
\item{readout}{Readout of the sliding window, e.g. number of integrations}
}
}
\description{
Given genomic coordinates (chromosome, start, end) a sliding window runs over
the genome with a specific window size and step size. Readouts of this window
are calculated with specific "readout" functions, e.g.
\code{\link{slidingWindowReadoutCounts}} returns the absolute number of
insertions within a window, however possibilities are countless.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
