% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findAmbiguousInsertionSites}
\alias{findAmbiguousInsertionSites}
\title{Find transposon insertion sites with ambiguous positions}
\usage{
findAmbiguousInsertionSites(dt, padding = 2)
}
\arguments{
\item{dt}{A data.table as returned by readPutativeInsertions.}

\item{padding}{Padding added around the TIS position that is used to evaluate
whether the next upstream or downstream read overlaps within a range of N bp.}
}
\value{
Returns read values for shifted insertions:
\describe{
\item{read_name}{Read name}
\item{chr}{Chromosome}
\item{TIS_start}{Insertion site start position}
\item{TIS_end}{Insertion site end position}
\item{strand}{Read strandedness}
\item{TIS_seq}{Insertion site overhang sequence}
}
}
\description{
Identify the reads for insertion sites where the exact insert position is
ambiguous due to locally shifted read positions.
}
\details{
Reads for which the previous or next read on the genome has shifted by
\emph{up to N bp} are detected by this function. Such reads should indicate
that despite the local shift, reads map to the identical insertion site.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
