% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromosomeSize}
\alias{getChromosomeSize}
\title{Get chromosome size}
\usage{
getChromosomeSize(chromosome, chromosomeSizesFile)
}
\arguments{
\item{chromosome}{Chromosome to return the size for, formatted as
"chr<number>".}

\item{chromosomeSizesFile}{Full path to chrom.sizes file.}
}
\value{
Returns integer-value of chromosome size.
}
\description{
Get chromosome size from a reference genomes' chrom.sizes file, e.g. from
UCSC or IGV.
}
\details{
Insertion site data are updated using a decision tree.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
