% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapInsertionSites.R
\name{mapInsertionSites}
\alias{mapInsertionSites}
\title{Map transposon insertion sites.}
\usage{
mapInsertionSites(
  dt,
  bam,
  overhang,
  depth = 1,
  gapWidth = 1,
  ignoreStrand = TRUE,
  samtoolsPath = NULL,
  ambiguousInsertions = NULL
)
}
\arguments{
\item{dt}{A data.table with TIS data taken directly from
\code{\link{readCalledInsertions}}.}

\item{bam}{Full path to BAM file(s).}

\item{overhang}{TIS overhang sequence, e.g. 'TTAA' for PiggyBac.}

\item{depth}{Minimal read depth to filter TISes by.}

\item{gapWidth}{Minimal permitted gap between reads for merging reads
into a single insertion site location.}

\item{ignoreStrand}{Boolean indicating whether to ignore strand direction
and merge reads nonetheless (TRUE) or to separate merging by strand
direction (FALSE)}

\item{samtoolsPath}{Full path to SAMTOOLS.}

\item{ambiguousInsertions}{A data table with ambiguous insertion site locations
taken directly from \code{\link{findAmbiguousInsertionSites}}.}
}
\value{
A data table with the columns:
\describe{
\item{chr, start, end}{Exact genomic position of the insertion site.}
\item{TIS_seq}{Overhang sequence of the TIS.}
\item{strand}{Strand on which the insertion has taken place.}
\item{region_start, region_end}{Exact genomic position of reads around TIS.}
\item{read_names}{Names of individual reads supporting a particular TIS.}
\item{revmap}{TO-DO}
\item{read_count}{Number of reads supporting a TIS.}
\item{mapq}{Median mapq of reads supporting a TIS.}
}
}
\description{
Maps transposon insertion sites (TIS) in high detail from locations called by
the pipeline.
}
\details{
Input data table \code{dt} of filtered R2 reads (MAPQ >= 10, proper pairs, no
PCR duplicates) where the transposon overhang sequence and position is
reported, are processed in this function. TISes can be filtered based on read
depth, ambiguous insertion positions or overhang sequences are corrected
for, and more detailed metadata per TIS are reported.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
