% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readBam}
\alias{readBam}
\title{Read BAM file(s)}
\usage{
readBam(bamFilePath, samtoolsPath, flags = 0)
}
\arguments{
\item{bamFilePath}{Full path(s) to BAM file(s).}

\item{samtoolsPath}{Full path to SAMTOOLS.}

\item{flags}{BAM flag(s) to filter reads by.}
}
\value{
Returns data frame with SAM-formatted data (see specification at
at \url{https://samtools.github.io/hts-specs/SAMv1.pdf}, however the default
QUAL column is removed here.
}
\description{
Function to read one or two BAM files with a Samtools command line call. The
data of two BAM files are combined into a single output.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
