% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCalledInsertions}
\alias{readCalledInsertions}
\title{Read insertion sites called by pipeline.}
\usage{
readCalledInsertions(inputFile)
}
\arguments{
\item{inputFile}{Path to input file.}
}
\value{
Returns data table for insertion site calling data outputted by the
TagMap pipeline.
\describe{
\item{read_name}{Read name}
\item{chr}{Read chromosome}
\item{read_start}{Read start position}
\item{read_end}{Read end position}
\item{mapq}{Read mapping quality (mapq)}
\item{strand}{Read strandedness}
\item{TIS_seq}{Insertion site overhang sequence}
\item{TIS_start}{Insertion site start position}
\item{TIS_end}{Insertion site end position}
\item{sample}{Experiment sample name, e.g. 'WT78_SB_Fw'}
}
}
\description{
Reads insertions called by the pipeline into a data.table, according to the
specific column naming convention, to be used for further analyses.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
