% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindowReadoutCounts}
\alias{slidingWindowReadoutCounts}
\title{Calculate the number of insertions within a sliding window}
\usage{
slidingWindowReadoutCounts(data, chromosome, windowStart, windowEnd)
}
\arguments{
\item{data}{Insertion site data (as formatted by TagMap pipeline), columns
should minimally include- and be named 'chr', 'start' and 'end.}

\item{chromosome}{Chromosome, formatted as "chr(number)".}

\item{windowStart}{Window start position.}

\item{windowEnd}{Window end position.}
}
\value{
Returns number of insertions in the subsetted data object.
}
\description{
Subsets the complete data object to only contain insertion site data within
the genomic coordinates of the sliding window.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
