% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindow.R
\name{slidingWindowReadoutFraction}
\alias{slidingWindowReadoutFraction}
\title{Calculate the fraction of insertions between experimental conditions
within a sliding window}
\usage{
slidingWindowReadoutFraction(
  data,
  chromosome,
  windowStart,
  windowEnd,
  numerator,
  denominator,
  pseudocount = 1
)
}
\arguments{
\item{data}{Insertion site data (as formatted by TagMap pipeline), columns
should minimally include and be formatted as 'chr', 'start', 'end' and
'class'.}

\item{chromosome}{Chromosome, formatted as "chr(number)".}

\item{windowStart}{Window start position.}

\item{windowEnd}{Window end position.}

\item{numerator}{Names of experimental condition(s) in column 'class' of data
to select numerator insertions for the fraction.}

\item{denominator}{of experimental condition(s) in column 'class' of
data to select denominator insertions for the fraction.}
}
\value{
Returns the fraction of insertions between experimental conditions.
}
\description{
Subsets the complete data object to only contain insertion site data within
the genomic coordinates of the sliding window. Next, insertions are divided
by experimental conditions ('class' column) into the numerator insertions and
denominator insertions (e.g. effect vs. control, respectively). The fraction
of insertions is calculated by dividing the numerator by the denominator, a
pseudocount of 1 is added for zero-values of the denominator.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
