% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateInsertionSiteData}
\alias{updateInsertionSiteData}
\title{Update insertion site data}
\usage{
updateInsertionSiteData(
  dt,
  gr,
  overhang,
  readsPerTIS,
  ambiguousInsertions = NULL
)
}
\arguments{
\item{dt}{A data table with either all or subsets of insertion site data, as
processed in the function mapInsertionSites.}

\item{gr}{A GRanges object with detailed paired-end read data underlying the
insertion sites in dt.}

\item{overhang}{Insertion site overhang sequence, e.g. 'TTAA' for PiggyBac.}

\item{readsPerTIS}{A data table of read counts and genomic coordinates per
unique insertion site, typically generated in the function mapInsertionSites.}

\item{ambiguousInsertions}{An optional data table with shifted insertion site
locations outputted by the function \code{\link{findAmbiguousInsertionSites}}
. When left out, only overhang sequence values are updated.}
}
\value{
Returns the input data table \code{dt} with updated values.
}
\description{
Update data for insertion sites, including the exact inserted sequence
("overhang") when read data is ambiguous and the exact position when read
data displays neighboring yet shifted reads (returned by
\code{\link{findAmbiguousInsertionSites}}).
}
\details{
Insertion site data are updated using a decision tree.

If argument \code{ambiguousInsertions} is passed, the ambiguous insertion
site positions are investigated: the position with highest number of reads is
returned if possible, otherwise the overhang sequences are compared to the
one expected from argument \code{overhang}. If all ambiguous positions match
the expected overhang, a random position will be returned, otherwise it is
checked if one position uniquely matches the overhang which will be returned
or that any positions matches the overhang at all, when none are present then
a random position will be returned and if more positions uniquely match the
overhang then a random position will be returned.

If no argument \code{ambiguousInsertions} is passed, insertion site overhang
sequences are investigated: if only one unique sequence is present, then this
will be returned, otherwise the sequence with the highest number of reads is
returned if possible and otherwise a random sequence is returned.
}
\author{
Koen Rademaker, \email{k.rademaker@nki.nl}
}
